adjusting socket services for extensibility
