+++ src/authentication.ts
@@ -6,8 +6,8 @@ import {
   AuthenticationStrategy,
   TokenService,
   StrutServices,
-  TokenResponse,
 } from './dataTypes';
+import { TokenResponse } from './socket/dataTypes';
 import { ModelData } from 'plump';
 import { StrutServer } from './dataTypes';
 
@@ -27,15 +27,16 @@ function routeGen(options: AuthenticationStrategy, strut: StrutServices) {
   );
   const routeHandler: Hapi.RouteHandler = (request, reply) => {
     return options.handler(request, strut).then(r => {
-      strut.io.to(request.state[`${options.name}-nonce`].nonce).emit(
-        request.state[`${options.name}-nonce`].nonce,
-        {
+      strut.io
+        .to(request.state[`${options.name}-nonce`].nonce)
+        .emit(request.state[`${options.name}-nonce`].nonce, {
           response: 'token',
           status: 'success',
           token: r.token,
-        } as TokenResponse,
-      );
-      reply(r.response).type('text/html').unstate(`${options.name}-nonce`);
+        } as TokenResponse);
+      reply(r.response)
+        .type('text/html')
+        .unstate(`${options.name}-nonce`);
     });
   };
   return server => {
