+++ src/authorize.ts
@@ -1,18 +1,180 @@
-import { Model, ModelData, ModelReference, Plump } from 'plump';
+import {
+  Model,
+  ModelData,
+  ModelReference,
+  Plump,
+  IndefiniteModelData,
+} from 'plump';
 import * as Hapi from 'hapi';
 import * as mergeOptions from 'merge-options';
 import * as Joi from 'joi';
 import * as Boom from 'boom';
 import {
-  AuthorizeRequest,
   SegmentGenerator,
   Transformer,
   RouteOptions,
   StrutRouteConfiguration,
   StrutServices,
-  ActorMapFn,
 } from './dataTypes';
 import { Oracle } from './oracle';
+
+export interface AbstractAuthorizeRequest {
+  kind: 'attributes' | 'relationship' | 'compound';
+}
+
+export interface AbstractAttributesAuthorizeRequest
+  extends AbstractAuthorizeRequest {
+  action: 'create' | 'read' | 'update' | 'delete' | 'query';
+  actor: ModelReference;
+  kind: 'attributes';
+}
+
+export interface AttributesReadAuthorizeRequest
+  extends AbstractAttributesAuthorizeRequest {
+  action: 'read';
+  target: ModelReference;
+}
+
+export interface AttributesDeleteAuthorizeRequest
+  extends AbstractAttributesAuthorizeRequest {
+  action: 'delete';
+  target: ModelReference;
+}
+
+export interface AttributesCreateAuthorizeRequest
+  extends AbstractAttributesAuthorizeRequest {
+  action: 'create';
+  data?: IndefiniteModelData;
+  target: {
+    type: string;
+  };
+}
+
+export interface AttributesQueryAuthorizeRequest
+  extends AbstractAttributesAuthorizeRequest {
+  action: 'query';
+  target: {
+    type: string;
+  };
+}
+
+export interface AttributesUpdateAuthorizeRequest
+  extends AbstractAttributesAuthorizeRequest {
+  action: 'update';
+  target: ModelReference;
+  data?: ModelData;
+}
+
+export type AttributesAuthorizeRequest =
+  | AttributesCreateAuthorizeRequest
+  | AttributesReadAuthorizeRequest
+  | AttributesUpdateAuthorizeRequest
+  | AttributesDeleteAuthorizeRequest
+  | AttributesQueryAuthorizeRequest;
+
+export interface AbstractRelationshipAuthorizeRequest
+  extends AbstractAuthorizeRequest {
+  kind: 'relationship';
+  actor: ModelReference;
+  action: 'create' | 'read' | 'update' | 'delete';
+  relationship: string;
+  target: ModelReference;
+}
+
+export interface RelationshipCreateAuthorizeRequest
+  extends AbstractRelationshipAuthorizeRequest {
+  action: 'create';
+  child: ModelReference;
+  meta?: any;
+}
+
+export interface RelationshipReadAuthorizeRequest
+  extends AbstractRelationshipAuthorizeRequest {
+  action: 'read';
+}
+
+export interface RelationshipUpdateAuthorizeRequest
+  extends AbstractRelationshipAuthorizeRequest {
+  action: 'update';
+  child: ModelReference;
+  meta?: any;
+}
+
+export interface RelationshipDeleteAuthorizeRequest
+  extends AbstractRelationshipAuthorizeRequest {
+  action: 'delete';
+  child: ModelReference;
+}
+
+export type RelationshipAuthorizeRequest =
+  | RelationshipCreateAuthorizeRequest
+  | RelationshipReadAuthorizeRequest
+  | RelationshipUpdateAuthorizeRequest
+  | RelationshipDeleteAuthorizeRequest;
+
+export type SimpleAuthorizeRequest =
+  | RelationshipAuthorizeRequest
+  | AttributesAuthorizeRequest;
+
+export interface CompoundAuthorizeRequest extends AbstractAuthorizeRequest {
+  kind: 'compound';
+  combinator: 'and' | 'or';
+  list: (
+    | AttributesAuthorizeRequest
+    | RelationshipAuthorizeRequest
+    | CompoundAuthorizeRequest)[];
+}
+
+export type ConcreteAuthorizeRequest =
+  | RelationshipAuthorizeRequest
+  | AttributesAuthorizeRequest;
+export type AuthorizeRequest =
+  | RelationshipAuthorizeRequest
+  | AttributesAuthorizeRequest
+  | CompoundAuthorizeRequest;
+
+export interface AbstractAuthorizeResponse {
+  kind: string;
+}
+
+export interface FinalAuthorizeResponse extends AbstractAuthorizeResponse {
+  kind: 'final';
+  result: boolean;
+}
+
+export interface DelegateAuthorizeResponse extends AbstractAuthorizeResponse {
+  kind: 'delegated';
+  delegate: AuthorizeRequest;
+}
+
+export type AuthorizeResponse =
+  | FinalAuthorizeResponse
+  | DelegateAuthorizeResponse;
+
+export interface ActorMapFn {
+  (m: ModelData): ModelReference;
+}
+
+export interface AuthorizerDefinition {
+  mapActor?: ActorMapFn;
+  authorize(req: AuthorizeRequest): Promise<AuthorizeResponse>;
+}
+
+export interface FilterDefinition {
+  type: 'white' | 'black';
+  attributes?: string[];
+  relationships?: string[];
+}
+
+export interface IOracle {
+  authorizers: { [name: string]: AuthorizerDefinition };
+  filters: { [name: string]: FilterDefinition };
+  addAuthorizer(auth: AuthorizerDefinition, forType: string): void;
+  addFilter(auth: FilterDefinition, forType: string): void;
+  filter(m: ModelData): ModelData;
+  dispatch(request: AuthorizeRequest): Promise<FinalAuthorizeResponse>;
+  authorize(request: AuthorizeRequest): Promise<boolean>;
+}
 
 function generateAuthRequest(
   options: RouteOptions,
